mroyt blink completed - questions embedded
questions in the comments:  /*Why do we need to use the "$" to make this
a jQuery element?
I'm having trouble understanding some of the logical flow
of the $ function combined with "this."*/

/*ALso, I originally tried using "this" with the fadeOut and fadeIn,
but it didn't work until I pulled it out and put it into the el
variable. Why is that?*/